Refactored views into new views/ modules, refs #256
